+++ cli-tool/app/commands/export.rb
@@ -27,9 +27,9 @@ module Taylor
           Options:
             --help\t\t\tDisplays this message
             --dry-run\t\t\tJust display the export command and don't run it
-            --export_directory directory\tWhat directory do you want your exports (defaults to ./exports)
-            --export_targets targets\tWhat exports do you want (defaults to linux,windows,osx,web)
-            --build_cache directory\tWhere do you want to store build cache (defaults to nil)
+            --export-directory directory\tWhat directory do you want your exports (defaults to ./exports)
+            --export-targets targets\tWhat exports do you want (defaults to linux,windows,osx,web)
+            --build-cache directory\tWhere do you want to store build cache (defaults to nil)
         STR
       end
 
@@ -45,15 +45,15 @@ module Taylor
       def setup_options(argv, options)
         parser = OptParser.new do |opts|
           opts.on(:help, :bool, false)
-          opts.on(:dry_run, :bool, false)
-          opts.on(:export_directory, :string, options.fetch("export_directory", "./exports"))
-          opts.on(:export_targets, :string, options.fetch("export_targets", "linux,windows,osx,web"))
-          opts.on(:build_cache, :string)
+          opts.on("dry-run", :bool, false)
+          opts.on("export-directory", :string, options.fetch("export_directory", "./exports"))
+          opts.on("export-targets", :string, options.fetch("export_targets", "linux,windows,osx/intel,osx/apple,web"))
+          opts.on("build-cache", :string)
         end
         parser.parse(argv, true)
 
         @options = parser.opts
-        @options[:export_targets] = @options[:export_targets].split(",") unless @options[:export_targets].is_a?(Array)
+        @options[:"export-targets"] = @options[:"export-targets"].split(",") unless @options[:"export-targets"].is_a?(Array)
       end
 
       def check_in_taylor_project!
@@ -63,40 +63,41 @@ module Taylor
       end
 
       def create_export_folder
-        return if @options[:dry_run]
+        return if @options[:"dry-run"]
 
-        unless File.exist?(options[:export_directory]) &&
-            File.directory?(options[:export_directory])
-          Dir.mkdir(options[:export_directory])
+        unless File.exist?(options[:"export-directory"]) &&
+            File.directory?(options[:"export-directory"])
+          Dir.mkdir(options[:"export-directory"])
         end
       end
 
       def create_build_cache_folder
-        return if @options[:dry_run]
-        return if @options[:build_cache].nil?
+        raise "Could not create folder as a file of the same name already exists" if File.exist?(options[:"build-cache"])
 
-        unless File.exist?(options[:build_cache]) &&
-            File.directory?(options[:build_cache])
-          Dir.mkdir(options[:build_cache])
-        end
+        return if @options[:"dry-run"]
+        return if @options[:"build-cache"].nil?
+        return if File.directory?(options[:"build-cache"])
+
+        Dir.mkdir(options[:"build-cache"])
       end
 
       def docker_build
         base_command = "docker run -u $(id -u ${USER}):$(id -g ${USER})"
         base_command << " --mount type=bind,source=#{Dir.pwd},target=/app/game"
-        base_command << " --mount type=bind,source=#{File.expand_path(@options[:export_directory])},target=/app/game/exports"
+        base_command << " --mount type=bind,source=#{File.expand_path(@options[:"export-directory"])},target=/app/game/exports"
 
-        unless @options[:build_cache].nil?
-          base_command << " --mount type=bind,source=#{@options[:build_cache]},target=/app/taylor/build/"
+        if @options[:"build-cache"]
+          create_build_cache_folder
+          base_command << " --mount type=bind,source=#{File.expand_path(@options[:"build-cache"])},target=/app/taylor/build/"
         end
 
-        @options[:export_targets].each do |target|
+        @options[:"export-targets"].each do |target|
           command = base_command.dup
 
           command << " --env EXPORT=#{target}" if target.include?("/")
           command << " hellrok/taylor:#{target.split("/").first}-v#{TAYLOR_VERSION}"
 
-          if options[:dry_run]
+          if options[:"dry-run"]
             puts command
           else
             `#{command}`