+++ cli-tool/test/commands/export_test.rb
@@ -13,70 +13,295 @@ class Test
         Dir.chdir("./test")
 
         assert_raise(RuntimeError) {
-          Taylor::Commands::Export.new(["--dry-run"], {})
+          Taylor::Commands::Export.new([], {})
         }
       ensure
         Dir.chdir(original_path)
       end
 
       def test_check_in_taylor_project_success
-        Taylor::Commands::Export.new(["--dry-run"], {})
+        Taylor::Commands::Export.new([], {})
         assert_true true
       rescue RuntimeError
         assert_true false
       end
 
-      def test_create_export_folder_dry_run
+      def test_create_export_directory_dry_run
         Taylor::Commands::Export.new(["--dry-run"], {})
         assert_false Dir.exist?("./exports")
       end
 
-      def test_create_build_cache_folder
-        Taylor::Commands::Export.new(["--dry-run"], {"build_cache" => "/tmp/build/"})
+      def test_create_export_directory_default
+        Taylor::Commands::Export.new([], {})
+        assert_true Dir.exist?("./exports")
+      ensure
+        Dir.rmdir("./exports")
+      end
+
+      def test_create_export_directory_options
+        Taylor::Commands::Export.new([], {"export_directory" => "releases_option"})
         assert_false Dir.exist?("./exports")
+        assert_true Dir.exist?("./releases_option")
+      ensure
+        Dir.rmdir("./releases_option")
       end
 
-      def test_check_docker_command_default_options
+      def test_create_export_directory_flag
+        Taylor::Commands::Export.new(["--export-directory", "releases_flag"], {})
+        assert_false Dir.exist?("./exports")
+        assert_true Dir.exist?("./releases_flag")
+      ensure
+        Dir.rmdir("./releases_flag")
+      end
+
+      def test_check_docker_command_default_dry_run
         export_command = Taylor::Commands::Export.new(["--dry-run"], {})
-        assert_include export_command.puts_data, "docker run"
-        assert_include export_command.puts_data, File.join(Dir.pwd, "exports")
-        assert_include export_command.puts_data, "hellrok/taylor:linux-v#{TAYLOR_VERSION}"
-        assert_include export_command.puts_data, "hellrok/taylor:windows-v#{TAYLOR_VERSION}"
-        assert_include export_command.puts_data, "hellrok/taylor:osx-v#{TAYLOR_VERSION}"
-        assert_include export_command.puts_data, "hellrok/taylor:web-v#{TAYLOR_VERSION}"
+        assert export_command.backtick_data.nil?
+
+        assert_include(
+          export_command.puts_data,
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "exports")},target=/app/game/exports",
+            "hellrok/taylor:linux-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+        assert_include(
+          export_command.puts_data,
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "exports")},target=/app/game/exports",
+            "hellrok/taylor:windows-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+        assert_include(
+          export_command.puts_data,
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "exports")},target=/app/game/exports",
+            "--env EXPORT=osx/intel",
+            "hellrok/taylor:osx-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+        assert_include(
+          export_command.puts_data,
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "exports")},target=/app/game/exports",
+            "--env EXPORT=osx/apple",
+            "hellrok/taylor:osx-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+        assert_include(
+          export_command.puts_data,
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "exports")},target=/app/game/exports",
+            "hellrok/taylor:web-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
       end
 
-      def test_check_docker_command_override_export_directory
-        export_command = Taylor::Commands::Export.new(["--dry-run"], {"export_directory" => "/tmp/exports"})
-        assert_include export_command.puts_data, "docker run"
-        assert_include export_command.puts_data, "/tmp/exports"
-        assert_include export_command.puts_data, "hellrok/taylor:linux-v#{TAYLOR_VERSION}"
-        assert_include export_command.puts_data, "hellrok/taylor:windows-v#{TAYLOR_VERSION}"
-        assert_include export_command.puts_data, "hellrok/taylor:osx-v#{TAYLOR_VERSION}"
-        assert_include export_command.puts_data, "hellrok/taylor:web-v#{TAYLOR_VERSION}"
+      def test_check_docker_command_default
+        export_command = Taylor::Commands::Export.new([], {})
+        assert export_command.puts_data.nil?
+
+        assert_equal(
+          export_command.backtick_data[0],
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "exports")},target=/app/game/exports",
+            "hellrok/taylor:linux-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+        assert_equal(
+          export_command.backtick_data[1],
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "exports")},target=/app/game/exports",
+            "hellrok/taylor:windows-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+        assert_equal(
+          export_command.backtick_data[2],
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "exports")},target=/app/game/exports",
+            "--env EXPORT=osx/intel",
+            "hellrok/taylor:osx-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+        assert_equal(
+          export_command.backtick_data[3],
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "exports")},target=/app/game/exports",
+            "--env EXPORT=osx/apple",
+            "hellrok/taylor:osx-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+        assert_equal(
+          export_command.backtick_data[4],
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "exports")},target=/app/game/exports",
+            "hellrok/taylor:web-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+      ensure
+        Dir.rmdir("exports")
       end
 
-      def test_check_docker_command_override_targets
-        export_command = Taylor::Commands::Export.new(["--dry-run"], {"export_targets" => "linux,web"})
-        assert_include export_command.puts_data, "docker run"
-        assert_include export_command.puts_data, File.join(Dir.pwd, "exports")
-        assert_include export_command.puts_data, "hellrok/taylor:linux-v#{TAYLOR_VERSION}"
-        assert_false export_command.puts_data.include? "hellrok/taylor:windows-v#{TAYLOR_VERSION}"
-        assert_false export_command.puts_data.include? "hellrok/taylor:osx-v#{TAYLOR_VERSION}"
-        assert_include export_command.puts_data, "hellrok/taylor:web-v#{TAYLOR_VERSION}"
+      def test_check_docker_command_export_directory_flag
+        export_command = Taylor::Commands::Export.new(["--export-directory", "releases_flag"], {})
+        skip "There's a bug in mruby-tiny-opt-parser"
+        assert_equal(
+          export_command.backtick_data[0],
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "releases_flag")},target=/app/game/exports",
+            "hellrok/taylor:linux-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+      ensure
+        begin
+          Dir.rmdir("releases_flag")
+        rescue
+          nil
+        end
+      end
+
+      def test_check_docker_command_export_directory_option
+        export_command = Taylor::Commands::Export.new([], {"export_directory" => "releases_option"})
+        assert_equal(
+          export_command.backtick_data[0],
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "releases_option")},target=/app/game/exports",
+            "hellrok/taylor:linux-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+      ensure
+        Dir.rmdir("releases_option")
+      end
+
+      def test_check_docker_command_export_targets_options
+        export_command = Taylor::Commands::Export.new([], {"export_targets" => "linux,web"})
+
+        assert_equal(export_command.backtick_data.size, 2)
+        assert_equal(
+          export_command.backtick_data[0],
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "exports")},target=/app/game/exports",
+            "hellrok/taylor:linux-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+        assert_equal(
+          export_command.backtick_data[1],
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "exports")},target=/app/game/exports",
+            "hellrok/taylor:web-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+      ensure
+        Dir.rmdir("exports")
+      end
+
+      def test_check_docker_command_export_targets_flag
+        export_command = Taylor::Commands::Export.new(["--export-targets", "linux,web"], {})
+        skip "There's a bug in mruby-tiny-opt-parser"
+
+        assert_equal(export_command.backtick_data.size, 2)
+        assert_equal(
+          export_command.backtick_data[0],
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "exports")},target=/app/game/exports",
+            "hellrok/taylor:linux-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+        assert_equal(
+          export_command.backtick_data[1],
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "exports")},target=/app/game/exports",
+            "hellrok/taylor:web-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+      ensure
+        begin
+          Dir.rmdir("exports")
+        rescue
+          nil
+        end
+
+        # This is due to the bug, can be removed once fixed
+        begin
+          Dir.rmdir("linux,web")
+        rescue
+          nil
+        end
+      end
+
+      def test_check_docker_command_set_build_cache_dry_run
+        export_command = Taylor::Commands::Export.new(["--dry-run", "--build-cache", "build_cache"], {})
+
+        assert_false Dir.exist?("./build_cache")
+        assert_true export_command.backtick_data.nil?
       end
 
       def test_check_docker_command_set_build_cache
-        # Because this isn't specified by the taylor-config.json we have to pass it
-        # in via the ARGV
-        export_command = Taylor::Commands::Export.new(["--dry-run", "--build_cache", "/tmp/build"], {})
-        assert_include export_command.puts_data, "docker run"
-        assert_include export_command.puts_data, File.join(Dir.pwd, "exports")
-        assert_include export_command.puts_data, "/tmp/build"
-        assert_include export_command.puts_data, "hellrok/taylor:linux-v#{TAYLOR_VERSION}"
-        assert_include export_command.puts_data, "hellrok/taylor:windows-v#{TAYLOR_VERSION}"
-        assert_include export_command.puts_data, "hellrok/taylor:osx-v#{TAYLOR_VERSION}"
-        assert_include export_command.puts_data, "hellrok/taylor:web-v#{TAYLOR_VERSION}"
+        export_command = Taylor::Commands::Export.new(["--build-cache", "build_cache_flag"], {})
+
+        assert_true Dir.exist?("./build_cache_flag")
+        assert_equal(
+          export_command.backtick_data[0],
+          [
+            "docker run",
+            "-u $(id -u ${USER}):$(id -g ${USER})",
+            "--mount type=bind,source=#{Dir.pwd},target=/app/game",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "exports")},target=/app/game/exports",
+            "--mount type=bind,source=#{File.join(Dir.pwd, "build_cache_flag")},target=/app/taylor/build/",
+            "hellrok/taylor:linux-v#{TAYLOR_VERSION}"
+          ].join(" ")
+        )
+      ensure
+        Dir.rmdir("build_cache_flag")
+        Dir.rmdir("exports")
       end
     end
   end